Move "getDisplayName" function to the correct file
